fix: 마이페이지 조회 기능을 MyPageController로 이동
